Scanner: remove dead scanner_task/scanner_tasks/scanner_queue plugin pipeline and move AddUser/AddAll/ProcessMedia/DeleteOldUserAlbums into api/scanner, leaving api/scanner/queue with only dispatch mechanics
